ft: Added python module hook to create custom reports. fix: Install dependencies now only operates when the environment is created. Now 'force_build' is a model parameter that can be placed in the config file to create the environment again and build src code. refac: Removed report methods from Experiment and ExperimentComparison and placed them as standalone functions. Moved Markdown class to report.py module. Moved ExperimentComparison to experiment.py module build: added scipy as standard dependency examples: added a custom report script to case example h
