+++ examples/case_g/custom_plot_script.py
