+++ floatcsep/utils.py
@@ -1,8 +1,6 @@
 # python libraries
 import copy
-import filecmp
 import functools
-import hashlib
 import itertools
 import logging
 import multiprocessing
@@ -23,7 +21,6 @@ import seaborn
 import shapely.geometry
 
 # pyCSEP libraries
-import six
 import yaml
 from csep.core.forecasts import GriddedForecast
 from csep.core.regions import CartesianGrid2D, compute_vertices
@@ -587,453 +584,10 @@ class TaskGraph:
         pass
 
 
-class MarkdownReport:
-    """Class to generate a Markdown report from a study."""
-
-    def __init__(self, outname="report.md"):
-        self.outname = outname
-        self.toc = []
-        self.has_title = True
-        self.has_introduction = False
-        self.markdown = []
-
-    def add_introduction(self, adict):
-        """Generate document header from dictionary."""
-        first = (
-            f"# CSEP Testing Results: {adict['simulation_name']}  \n"
-            f"**Forecast Name:** {adict['forecast_name']}  \n"
-            f"**Simulation Start Time:** {adict['origin_time']}  \n"
-            f"**Evaluation Time:** {adict['evaluation_time']}  \n"
-            f"**Catalog Source:** {adict['catalog_source']}  \n"
-            f"**Number Simulations:** {adict['num_simulations']}\n"
-        )
-
-        # used to determine to place TOC at beginning of document or after
-        # introduction.
-
-        self.has_introduction = True
-        self.markdown.append(first)
-        return first
-
-    def add_text(self, text):
-        """
-        Text should be a list of strings where each string will be on its own.
-
-        line. Each add_text command represents a paragraph.
-
-        Args:
-            text (list): lines to write
-        Returns:
-        """
-        self.markdown.append("  ".join(text) + "\n\n")
-
-    def add_figure(
-        self,
-        title,
-        relative_filepaths,
-        level=2,
-        ncols=1,
-        add_ext=False,
-        text="",
-        caption="",
-        width=None,
-    ):
-        """
-        This function expects a list of filepaths.
-
-        If you want the output
-        stacked, select a value of ncols. ncols should be divisible by
-        filepaths. todo: modify formatted_paths to work when not divis.
-
-        Args:
-            width:
-            caption:
-            text:
-            add_ext:
-            ncols:
-            title: name of the figure
-            level (int): value 1-6 depending on the heading
-            relative_filepaths (str or List[Tuple[str]]): list of paths in
-                order to make table
-        Returns:
-        """
-        # verify filepaths have proper extension should always be png
-        is_single = False
-        paths = []
-        if isinstance(relative_filepaths, six.string_types):
-            is_single = True
-            paths.append(relative_filepaths)
-        else:
-            paths = relative_filepaths
-
-        correct_paths = []
-        if add_ext:
-            for fp in paths:
-                correct_paths.append(fp + ".png")
-        else:
-            correct_paths = paths
-
-        # generate new lists with size ncols
-        formatted_paths = [correct_paths[i : i + ncols] for i in range(0, len(paths), ncols)]
-
-        # convert str into a list, where each potential row is an iter not str
-        def build_header(_row):
-            top = "|"
-            bottom = "|"
-            for i, _ in enumerate(_row):
-                if i == ncols:
-                    break
-                top += " |"
-                bottom += " --- |"
-            return top + "\n" + bottom
-
-        size_ = bool(width) * f"width={width}"
-
-        def add_to_row(_row):
-            if len(_row) == 1:
-                return f'<img src="{_row[0]}" {size_}/>'
-            string = "| "
-            for item in _row:
-                string = string + f'<img src="{item}" width={width}/>'
-            return string
-
-        level_string = f"{level * '#'}"
-        result_cell = []
-        locator = title.lower().replace(" ", "_")
-        result_cell.append(f'{level_string} {title}  <a name="{locator}"></a>\n')
-        result_cell.append(f"{text}\n")
-
-        for i, row in enumerate(formatted_paths):
-            if i == 0 and not is_single and ncols > 1:
-                result_cell.append(build_header(row))
-            result_cell.append(add_to_row(row))
-        result_cell.append("\n")
-        result_cell.append(f"{caption}")
-
-        self.markdown.append("\n".join(result_cell) + "\n")
-
-        # generate metadata for TOC
-        self.toc.append((title, level, locator))
-
-    def add_heading(self, title, level=1, text="", add_toc=True):
-        # multiplying char simply repeats it
-        if isinstance(text, str):
-            text = [text]
-        cell = []
-        level_string = f"{level * '#'}"
-        locator = title.lower().replace(" ", "_")
-        sub_heading = f'{level_string} {title} <a name="{locator}"></a>\n'
-        cell.append(sub_heading)
-        try:
-            for item in list(text):
-                cell.append(item)
-        except Exception as ex:
-            raise RuntimeWarning("Unable to add document subhead, text must be iterable.")
-        self.markdown.append("\n".join(cell) + "\n")
-
-        # generate metadata for TOC
-        if add_toc:
-            self.toc.append((title, level, locator))
-
-    def add_list(self, _list):
-        cell = []
-        for item in _list:
-            cell.append(f"* {item}")
-        self.markdown.append("\n".join(cell) + "\n\n")
-
-    def add_title(self, title, text):
-        self.has_title = True
-        self.add_heading(title, 1, text, add_toc=False)
-
-    def table_of_contents(self):
-        """Generates table of contents based on contents of document."""
-        if len(self.toc) == 0:
-            return
-        toc = ["# Table of Contents"]
-
-        for i, elem in enumerate(self.toc):
-            title, level, locator = elem
-            space = "   " * (level - 1)
-            toc.append(f"{space}1. [{title}](#{locator})")
-        insert_loc = 1 if self.has_title else 0
-        self.markdown.insert(insert_loc, "\n".join(toc) + "\n\n")
-
-    def add_table(self, data, use_header=True):
-        """
-        Generates table from HTML and styles using bootstrap class.
-
-        Args:
-           data List[Tuple[str]]: should be (nrows, ncols) in size. all rows
-            should be the same sizes
-        Returns:
-            table (str): this can be added to subheading or other cell if
-                desired.
-        """
-        table = ['<div class="table table-striped">', f"<table>"]
-
-        def make_header(row):
-            header = ["<tr>"]
-            for item in row:
-                header.append(f"<th>{item}</th>")
-            header.append("</tr>")
-            return "\n".join(header)
-
-        def add_row(row):
-            table_row = ["<tr>"]
-            for item in row:
-                table_row.append(f"<td>{item}</td>")
-            table_row.append("</tr>")
-            return "\n".join(table_row)
-
-        for i, row in enumerate(data):
-            if i == 0 and use_header:
-                table.append(make_header(row))
-            else:
-                table.append(add_row(row))
-        table.append("</table>")
-        table.append("</div>")
-        table = "\n".join(table)
-        self.markdown.append(table + "\n\n")
-
-    def save(self, save_dir):
-        output = list(itertools.chain.from_iterable(self.markdown))
-        full_md_fname = os.path.join(save_dir, self.outname)
-        with open(full_md_fname, "w") as f:
-            f.writelines(output)
-
-
 class NoAliasLoader(yaml.Loader):
     @staticmethod
     def ignore_aliases(self):
         return True
-
-
-class ExperimentComparison:
-
-    def __init__(self, original, reproduced, **kwargs):
-        """"""
-        self.original = original
-        self.reproduced = reproduced
-
-        self.num_results = {}
-        self.file_comp = {}
-
-    @staticmethod
-    def obs_diff(obs_orig, obs_repr):
-
-        return numpy.abs(
-            numpy.divide((numpy.array(obs_orig) - numpy.array(obs_repr)), numpy.array(obs_orig))
-        )
-
-    @staticmethod
-    def test_stat(test_orig, test_repr):
-
-        if isinstance(test_orig[0], str):
-            if not isinstance(test_orig[1], str):
-                stats = numpy.array(
-                    [0, numpy.divide((test_repr[1] - test_orig[1]), test_orig[1]), 0, 0]
-                )
-            else:
-                stats = None
-        else:
-            stats_orig = numpy.array(
-                [numpy.mean(test_orig), numpy.std(test_orig), scipy.stats.skew(test_orig)]
-            )
-            stats_repr = numpy.array(
-                [numpy.mean(test_repr), numpy.std(test_repr), scipy.stats.skew(test_repr)]
-            )
-
-            ks = scipy.stats.ks_2samp(test_orig, test_repr)
-            stats = [*numpy.divide(numpy.abs(stats_repr - stats_orig), stats_orig), ks.pvalue]
-        return stats
-
-    def get_results(self):
-
-        win_orig = timewindow2str(self.original.timewindows)
-        win_repr = timewindow2str(self.reproduced.timewindows)
-
-        tests_orig = self.original.tests
-        tests_repr = self.reproduced.tests
-
-        models_orig = [i.name for i in self.original.models]
-        models_repr = [i.name for i in self.reproduced.models]
-
-        results = dict.fromkeys([i.name for i in tests_orig])
-
-        for test in tests_orig:
-            if test.type in ["consistency", "comparative"]:
-                results[test.name] = dict.fromkeys(win_orig)
-                for tw in win_orig:
-                    results_orig = self.original.read_results(test, tw)
-                    results_repr = self.reproduced.read_results(test, tw)
-                    results[test.name][tw] = {
-                        models_orig[i]: {
-                            "observed_statistic": self.obs_diff(
-                                results_orig[i].observed_statistic,
-                                results_repr[i].observed_statistic,
-                            ),
-                            "test_statistic": self.test_stat(
-                                results_orig[i].test_distribution,
-                                results_repr[i].test_distribution,
-                            ),
-                        }
-                        for i in range(len(models_orig))
-                    }
-
-            else:
-                results_orig = self.original.read_results(test, win_orig[-1])
-                results_repr = self.reproduced.read_results(test, win_orig[-1])
-                results[test.name] = {
-                    models_orig[i]: {
-                        "observed_statistic": self.obs_diff(
-                            results_orig[i].observed_statistic,
-                            results_repr[i].observed_statistic,
-                        ),
-                        "test_statistic": self.test_stat(
-                            results_orig[i].test_distribution, results_repr[i].test_distribution
-                        ),
-                    }
-                    for i in range(len(models_orig))
-                }
-
-        return results
-
-    @staticmethod
-    def get_hash(filename):
-
-        with open(filename, "rb") as f:
-            bytes_file = f.read()
-            readable_hash = hashlib.sha256(bytes_file).hexdigest()
-        return readable_hash
-
-    def get_filecomp(self):
-
-        win_orig = timewindow2str(self.original.timewindows)
-        win_repr = timewindow2str(self.reproduced.timewindows)
-
-        tests_orig = self.original.tests
-        tests_repr = self.reproduced.tests
-
-        models_orig = [i.name for i in self.original.models]
-        models_repr = [i.name for i in self.reproduced.models]
-
-        results = dict.fromkeys([i.name for i in tests_orig])
-
-        for test in tests_orig:
-            if test.type in ["consistency", "comparative"]:
-                results[test.name] = dict.fromkeys(win_orig)
-                for tw in win_orig:
-                    results[test.name][tw] = dict.fromkeys(models_orig)
-                    for model in models_orig:
-                        orig_path = self.original.registry.get_result(tw, test, model)
-                        repr_path = self.reproduced.registry.get_result(tw, test, model)
-
-                        results[test.name][tw][model] = {
-                            "hash": (self.get_hash(orig_path) == self.get_hash(repr_path)),
-                            "byte2byte": filecmp.cmp(orig_path, repr_path),
-                        }
-            else:
-                results[test.name] = dict.fromkeys(models_orig)
-                for model in models_orig:
-                    orig_path = self.original.registry.get_result(win_orig[-1], test, model)
-                    repr_path = self.reproduced.registry.get_result(win_orig[-1], test, model)
-                    results[test.name][model] = {
-                        "hash": (self.get_hash(orig_path) == self.get_hash(repr_path)),
-                        "byte2byte": filecmp.cmp(orig_path, repr_path),
-                    }
-        return results
-
-    def compare_results(self):
-
-        self.num_results = self.get_results()
-        self.file_comp = self.get_filecomp()
-        self.write_report()
-
-    def write_report(self):
-
-        numerical = self.num_results
-        data = self.file_comp
-        outname = os.path.join("reproducibility_report.md")
-        save_path = os.path.dirname(
-            os.path.join(self.reproduced.registry.workdir, self.reproduced.registry.run_dir)
-        )
-        report = MarkdownReport(outname=outname)
-        report.add_title(f"Reproducibility Report - {self.original.name}", "")
-
-        report.add_heading("Objectives", level=2)
-        objs = [
-            "Analyze the statistic reproducibility and data reproducibility of"
-            " the experiment. Compares the differences between "
-            "(i) the original and reproduced scores,"
-            " (ii) the statistical descriptors of the test distributions,"
-            " (iii) The p-value of a Kolmogorov-Smirnov test -"
-            " values beneath 0.1 means we can't reject the distributions are"
-            " similar -,"
-            " (iv) Hash (SHA-256) comparison between the results' files and "
-            "(v) byte-to-byte comparison"
-        ]
-
-        report.add_list(objs)
-        for num, dat in zip(numerical.items(), data.items()):
-
-            res_keys = list(num[1].keys())
-            is_time = False
-            try:
-                str2timewindow(res_keys[0])
-                is_time = True
-            except ValueError:
-                pass
-            if is_time:
-                report.add_heading(num[0], level=2)
-                for tw in res_keys:
-                    rows = [
-                        [
-                            tw,
-                            "Score difference",
-                            "Test Mean  diff.",
-                            "Test Std  diff.",
-                            "Test Skew  diff.",
-                            "KS-test p value",
-                            "Hash (SHA-256) equal",
-                            "Byte-to-byte equal",
-                        ]
-                    ]
-
-                    for model_stat, model_file in zip(num[1][tw].items(), dat[1][tw].items()):
-                        obs = model_stat[1]["observed_statistic"]
-                        test = model_stat[1]["test_statistic"]
-                        rows.append(
-                            [
-                                model_stat[0],
-                                obs,
-                                *[f"{i:.1e}" for i in test[:-1]],
-                                f"{test[-1]:.1e}",
-                                model_file[1]["hash"],
-                                model_file[1]["byte2byte"],
-                            ]
-                        )
-                    report.add_table(rows)
-            else:
-                report.add_heading(num[0], level=2)
-                rows = [
-                    [tw, "Max Score difference", "Hash (SHA-256) equal", "Byte-to-byte equal"]
-                ]
-
-                for model_stat, model_file in zip(num[1].items(), dat[1].items()):
-                    obs = numpy.nanmax(model_stat[1]["observed_statistic"])
-
-                    rows.append(
-                        [
-                            model_stat[0],
-                            f"{obs:.1e}",
-                            model_file[1]["hash"],
-                            model_file[1]["byte2byte"],
-                        ]
-                    )
-
-                report.add_table(rows)
-        report.table_of_contents()
-        report.save(save_path)
 
 
 #######################
