+++ setup.cfg
@@ -33,6 +33,7 @@ install_requires =
     pyshp
     pyyaml
     requests
+    scipy
     seaborn
     tables
     xmltodict
